Split out CellIterator
